requirements.txt
streamlit>=1.40.1,<2.0
insightface==0.7.3
onnxruntime==1.18.0
opencv-python-headless>=4.10,<5     # ← updated
numpy==1.26.4
pillow==10.3.0
scipy==1.13.1
